+++ packages/interceptable/.eslintrc.js
@@ -5,4 +5,12 @@ module.exports = {
 	$schema: "http://json.schemastore.org/eslintrc",
 	root: true,
 	extends: ["custom"],
+	overrides: [
+		{
+			files: ["*test.ts"],
+			env: {
+				node: true,
+			},
+		},
+	],
 };