+++ packages/resize-observer-store/.eslintrc.js
@@ -1,3 +1,6 @@
+/**
+ * @type {import("eslint").Linter.Config}
+ */
 module.exports = {
 	$schema: "http://json.schemastore.org/eslintrc",
 	root: true,